Upgrade Node.js from 10.x to 12.x
In preparation for migrating to 20.x we need to make
an incremental upgrade.

This is because Node.js 20.x does not run on Centos 7.
However using our up-to-date Node.js 20 build (ci-node-build-20)
Ruby 2.3 fails to build.

The CDN requires Ruby (Ruby Sass) to build .scss files into .scss files,
but we have done work to remove this requirement
allowing the use of Dart Sass which only depends on Node.js.

Our Dart Sass improvements do not run on Node.js 10.x, they run at minimum
on Node.js 12.x

So by upgrading this image to Node.js 12.x we allow
this image to build the newer Dart Sass and Ruby Sass versions of the CDN.

After we confirmed this is working we would then remove the Ruby Sass dependency
in the CDN, which would allow us to drastically simplify this image and move away from
this out of date image entirely but it means we can roll back if we need to.